Move page code to src/pages
