+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
-import './index.scss'
-import { App } from './App'
+import './styles/index.scss'
+import { App } from './pages/App'
 import { reportWebVitals } from './reportWebVitals'
 
 ReactDOM.render(
